refactoring seed file
