+++ src/seed.ts
@@ -1,50 +1,10 @@
 import { faker } from "@faker-js/faker";
-import { prisma } from "./setup-db";
-import { hashPassword } from "./graphql/resolvers";
+import { createRandomUser, userSeed, userSeedInfo } from "./user-seed";
+import { setupDatabase } from "./setup-db";
 
-interface userSeedInfo {
-  name: string;
-  email: string;
-  password: string;
-  birthDate: string;
-}
-function createRandomUser(): userSeedInfo {
-  return {
-    password: faker.internet.password(),
-    email: faker.internet.email(),
-    name: faker.internet.userName(),
-    birthDate: faker.date
-      .birthdate()
-      .toLocaleDateString("en-GB", { day: "2-digit", month: "2-digit", year: "numeric" })
-      .replace(/\//g, "-")
-  };
+async function populateDatabase(limit: number) {
+  await setupDatabase();
+  await userSeed(limit);
 }
 
-const USERS: userSeedInfo[] = faker.helpers.multiple(createRandomUser, {
-  count: 5
-});
-
-async function seed() {
-  const usersWithHashedPasswords = await Promise.all(
-    USERS.map(async (user) => {
-      const hashedPassword = await hashPassword(user.password);
-      return {
-        ...user,
-        password: hashedPassword
-      };
-    })
-  );
-
-  await prisma.user.createMany({
-    data: usersWithHashedPasswords,
-    skipDuplicates: true
-  });
-}
-
-export async function populateDatabase(limit?: number) {
-  const USERS: userSeedInfo[] = faker.helpers.multiple(createRandomUser, {
-    count: limit ?? 5
-  });
-
-  await seed(USERS);
-}
+populateDatabase(15);
